Move BasketCache to new package
Prep work for unifying the various caches, move cache out of the
top-level laurelin package to its own sub-package
