+++ src/main/java/edu/vanderbilt/accre/laurelin/cache/Cache.java
@@ -1,4 +1,4 @@
-package edu.vanderbilt.accre.laurelin;
+package edu.vanderbilt.accre.laurelin.cache;
 
 import java.lang.ref.SoftReference;
 import java.util.Collections;
