+++ src/main/java/edu/vanderbilt/accre/laurelin/cache/CacheFactory.java
@@ -1,4 +1,4 @@
-package edu.vanderbilt.accre.laurelin;
+package edu.vanderbilt.accre.laurelin.cache;
 
 import java.io.Serializable;
