+++ src/main/java/edu/vanderbilt/accre/laurelin/cache/CacheStash.java
@@ -1,4 +1,4 @@
-package edu.vanderbilt.accre.laurelin;
+package edu.vanderbilt.accre.laurelin.cache;
 
 public class CacheStash {
     private static Cache cache;
