+++ src/main/java/edu/vanderbilt/accre/laurelin/spark_ttree/SlimTBranchInterface.java
@@ -2,8 +2,8 @@ package edu.vanderbilt.accre.laurelin.spark_ttree;
 
 import com.google.common.collect.ImmutableRangeMap;
 
-import edu.vanderbilt.accre.laurelin.Cache;
 import edu.vanderbilt.accre.laurelin.array.ArrayBuilder;
+import edu.vanderbilt.accre.laurelin.cache.Cache;
 import edu.vanderbilt.accre.laurelin.root_proxy.ROOTFileCache;
 import edu.vanderbilt.accre.laurelin.root_proxy.TBranch;
 import edu.vanderbilt.accre.laurelin.spark_ttree.SlimTBranch.SlimTBasket;
